+++ substrate/state-machine/src/ext.rs
@@ -154,7 +154,7 @@ where
 	T: 'a + ChangesTrieStorage<H>,
 	H::Out: Ord + Encodable + HeapSizeOf,
 {
-	fn set_changes_trie_config(&mut self, block: u64, digest_interval: u64, digest_levels: u8) {
+	fn set_changes_trie_config(&mut self, block: u64, digest_interval: u64, digest_levels: u32) {
 		self.overlay.set_changes_trie_config(block, ChangesTrieConfig {
 			digest_interval,
 			digest_levels,
